requirements/base.txt
asgiref==3.4.1
Django==3.2
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
isort==5.9.3
black
flake8==3.9.2
mypy==0.910
django-stubs==1.8.0
djangorestframework-stubs==1.4.0